Fix coprocessor unusable accuracy (#88)
* Add floating point registers views in the debugger to see the value of each registers interpreted as S, D, W or L

* Add COP2 opcode. This is used by systemtest to initilize the exception registers to a known state. At this stage I'm not sure if a generic COP2 opcode is enough or if we need to implement generic coprocessor opcodes such as MFC2, MTC2 etc..

* Support coprocessor 2 in coprocessor unusable exception.

* Implement Floating point exception and fpe triggered through ctc1 opcode.

* LogMessage buffer should be large enough for ISViewer lines

* Partial implementation of FPU exceptions for ADD.S opcode.

* Fix all flags, exceptions etc.. detected by n64-systemtest for COP1.S opcode in interpreter.

* Fix exceptions and flags of ADD.D opcodes in interpreter. Hopefully, the other cop1 computation opcodes can be written using the same template.

* SUB.S is passing n64-systemtest.

* SUB.D passes n64-systemtest in interpreter

* MUL.S is passing systemtest in interpreter.

* MUL.D is passing systemtest in interpreter.

* DIV.S and DIV.D passes systemtest in interpreter.

* ABS.S and ABS.D pass systemtest in interpreter.

* Fix NEG.S, NEG.D, SQRT.S, SQRT.D n64-systemtest tests in interpreter.

* Fix Index register mask in interpreter.

* Fix CVT.S.S, CVT.S.D, CVT.S.W, CVT.S.L n64-systemtest in interpreter.

* Fix CVT.D.S, CVT.D.D, CVT.D.W, CVT.D.L tests in n64-systemtest in interpreter.

* CVT.W.S, ROUND.W.S, CEIL.W.S, FLOOR.W.S, TRUNC.W.S passes n64-systemtest in interpreter
CVT.W.W, CVT.W.L added.

* Fix CVT.W.D, ROUND.W.D, TRUNC.W.D, CEIL.W.D, FLOOW.W.D n64-systemtest tests in interpreter.
CVT.W.W, CVT.W.L generate unimplmented FPE
Add ROUND.W.W, TRUNC.W.W, CEIL.W.W, FLOOR.W.W, ROUND.W.L, TRUNC.W.L, CEIL.W.L, FLOOR.W.L, they all generate unimplemented FPE.

* Add CVT.L.W and CVT.L.L
Fix n64-systemtest tests in interpreter for CVT.L.S, ROUND.L.S, TRUNC.L.S, CEIL.L.S and FLOOR.L.S

* Add in interpreter CVT.L.W, ROUND.L.W, TRUNC.L.W, CEIL.L.W, FLOOR.L.W, CVT.L.L, ROUND.L.L, TRUNC.L.L, CEIL.L.L, FLOOR.L.L (they generate unimplemented FPE).
Fix n64-systemtest tests in interpreter for CVT.L.S, ROUND.L.S, TRUNC.L.S, CEIL.L.S, FLOOR.L.S

* Fix all FPU comparison opcodes in interpreter. It's passing all fpu tests in n64-systemtest now.

* sign extend EPC register in cop unusable exception

* Add DCFC1 (generate unimplmented fpe)

* Add DCTC1 (generate unimplemented FPE) in interpreter.
Add DCFC1 and DCTC1 in debugger.

* * split cop2 into MFC2, DMFC2, CFC2, DCFC2, MTC2, DMTC2, CTC2
* fix exceptions generated by DCFC2

* Add DCTC2 in interpreter.

* Add COP3 in interpreter.

* Implement MFC2,DMFC2,MTC2,DMTC2 in interpreter.

+++ Exception.c
@@ -169,9 +169,9 @@ void _fastcall DoCopUnusableException ( BOOL DelaySlot, int Coprocessor ) {
 	else if (Coprocessor == 2) { CAUSE_REGISTER |= 0x20000000; }
 	if (DelaySlot) {
 		CAUSE_REGISTER |= CAUSE_BD;
-		EPC_REGISTER = PROGRAM_COUNTER - 4;
+		EPC_REGISTER = (long)(PROGRAM_COUNTER - 4);
 	} else {
-		EPC_REGISTER = PROGRAM_COUNTER;
+		EPC_REGISTER = (long)PROGRAM_COUNTER;
 	}
 	STATUS_REGISTER |= STATUS_EXL;
 	PROGRAM_COUNTER = 0x80000180;
@@ -254,10 +254,10 @@ void _fastcall DoIllegalInstructionException(BOOL DelaySlot) {
 	CAUSE_REGISTER = EXC_II;
 	if (DelaySlot) {
 		CAUSE_REGISTER |= CAUSE_BD;
-		EPC_REGISTER = PROGRAM_COUNTER - 4;
+		EPC_REGISTER = (long)(PROGRAM_COUNTER - 4);
 	}
 	else {
-		EPC_REGISTER = PROGRAM_COUNTER;
+		EPC_REGISTER = (long)PROGRAM_COUNTER;
 	}
 	STATUS_REGISTER |= STATUS_EXL;
 	PROGRAM_COUNTER = 0x80000180;
